Move About related activities/fragments to about package
