+++ app/src/androidTest/java/org/gdg/frisbee/android/activity/AboutActivityTest.java
@@ -19,6 +19,7 @@ package org.gdg.frisbee.android.activity;
 import android.test.ActivityInstrumentationTestCase2;
 
 import org.gdg.frisbee.android.R;
+import org.gdg.frisbee.android.about.AboutActivity;
 
 import static android.support.test.espresso.Espresso.onView;
 import static android.support.test.espresso.assertion.ViewAssertions.matches;
