+++ app/src/main/java/org/gdg/frisbee/android/activity/GdgNavDrawerActivity.java
@@ -43,6 +43,7 @@ import com.google.api.services.plus.model.Person;
 import org.gdg.frisbee.android.BuildConfig;
 import org.gdg.frisbee.android.Const;
 import org.gdg.frisbee.android.R;
+import org.gdg.frisbee.android.about.AboutActivity;
 import org.gdg.frisbee.android.adapter.DrawerAdapter;
 import org.gdg.frisbee.android.adapter.DrawerAdapter.DrawerItem;
 import org.gdg.frisbee.android.api.GapiOkTransport;
